Updated get_pair_params()
Fixed: OLS minimises the sum of squared residuals but doesn't necessarily nullify the sum of residuals, meaning that the mean of the spread will rarely be exactly 0. So it's actually more accurate to calculate the number of zero crossings of the Z-score of the spread. Edited get_pair_params() so it does just that.